Optimization changes (details in the description)
- Additions: if you try to insert an email in the monitor form without a
"@" or "." an error message pops up.
- Button actions when pressing "Enter" key: ventanaLogin, form
Activities, form Monitors, form Accommodations, form Routes &
configuration.
- Inconsistencies: between the list and the form of the monitors (the
phone number was not in the same format), when you entered in
ventanaInicio the button "Actividades" was not highlighted & when
changing the language to English the colors of the availability of
monitors were not changing.
- Internationalization: the error message that says the email does not
have "@" or "." in the monitor form & reservation cancelling messages.
- Typos: (English) while inserting a new language it needed a space & in
some parts "Accommodation" was written with just one "m"